backend/pyproject.toml
[project]
name = "gitm8"
version = "0.1.0"
description = ""
authors = [
    {name = "Davshiv20",email = "shivamdave2903@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "aiohttp (>=3.8.0,<4.0.0)",
    "pydantic (>=2.0.0,<3.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "google-genai (>=1.28.0,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
